Cleaned up files (moved feature grid etc to ViewFeatures), for a less modified ViewDatasets component.
