+++ ui/client/components/ViewDatasets.js
@@ -1,9 +1,6 @@
 import React, { useEffect, useState, useCallback } from 'react';
-
 import { Link } from 'react-router-dom';
-
 import axios from 'axios';
-import debounce from 'lodash/debounce';
 
 import Button from '@material-ui/core/Button';
 import ToggleButton from '@material-ui/lab/ToggleButton';
@@ -11,27 +8,21 @@ import ToggleButtonGroup from '@material-ui/lab/ToggleButtonGroup';
 import Container from '@material-ui/core/Container';
 import { GridOverlay, DataGrid, useGridSlotComponentProps } from '@material-ui/data-grid';
 import Typography from '@material-ui/core/Typography';
-import { darken, makeStyles, withStyles } from '@material-ui/core/styles';
-import Tooltip from '@material-ui/core/Tooltip';
-import TablePagination from '@material-ui/core/TablePagination';
-import Alert from '@material-ui/lab/Alert';
-
+import { darken, makeStyles } from '@material-ui/core/styles';
 import ExpandableDataGridCell from './ExpandableDataGridCell';
 import LoadingOverlay from './LoadingOverlay';
 import SearchDatasets from './SearchDatasets';
 
-import LinearProgress from '@material-ui/core/LinearProgress';
-import CancelIcon from '@material-ui/icons/Cancel';
-import OpenInNewIcon from '@material-ui/icons/OpenInNew';
-import IconButton from '@material-ui/core/IconButton';
-import InputAdornment from '@material-ui/core/InputAdornment';
-import TextField from '@material-ui/core/TextField';
+import ViewFeatures from './ViewFeatures';
+
 
 const useStyles = makeStyles((theme) => ({
   root: {
     padding: [[theme.spacing(8), theme.spacing(2), theme.spacing(2)]],
   },
   gridContainer: {
+    display: "flex",
+    flexDirection: "column",
     flex: 1,
     maxWidth: '2000px',
     minWidth: '900px',
@@ -68,8 +59,6 @@ const useStyles = makeStyles((theme) => ({
     display: 'flex',
     justifyContent: 'space-between',
     alignItems: 'center',
-  },
-  featureGridRoot: {
   }
 }));
 
@@ -146,54 +135,6 @@ const columns = [
     },
   ];
 
-const featureColumns = [
-    {
-      field: 'name',
-      headerName: 'Name',
-      minWidth: 200,
-      flex: 1
-    },
-    {
-      field: 'display_name',
-      headerName: 'Display Name',
-      renderCell: expandableCell,
-      minWidth: 200,
-      flex: 1
-    },
-    {
-      field: 'description',
-      headerName: 'Description',
-      renderCell: expandableCell,
-      minWidth: 200,
-      flex: 1
-    },
-    {
-      field: 'owner_dataset_name',
-      headerName: 'Dataset Name',
-      valueGetter: (cell) => cell.row.owner_dataset.name,
-      minWidth: 200,
-      flex: 1
-    },
-    {
-      field: 'link',
-      headerName: ' ',
-      sortable: false,
-      disableColumnMenu: true,
-      renderCell: ({ row }) => (
-        <Button
-          component="a"
-          href={`/dataset_summary?dataset=${row.owner_dataset.id}`}
-          target="_blank"
-          variant="outlined"
-        >
-          Parent Dataset
-          <OpenInNewIcon />
-        </Button>
-      ),
-      minWidth: 210,
-    }
-  ];
-
 const getDatasets = (setDatasets, setDatasetsError, setDatasetsLoading) => {
   // only do this for the first call to getDatasets, when we don't have a scrollId
   // so we don't show the full page spinner for every subsequent set of models
@@ -215,99 +156,7 @@ const getDatasets = (setDatasets, setDatasetsError, setDatasetsLoading) => {
     });
 };
 
-/**
- * Adapted from ViewModels.js::fetchModels
- */
-const fetchFeatures = async (
-  setFeatures, setFeaturesLoading, setFeaturesError, searchTerm, scrollId
-) => {
-  setFeaturesLoading(true);
 
-  let url = `/api/dojo/features`;
-  if (scrollId) {
-    url += `?scroll_id=${scrollId}`;
-  } else if (searchTerm) {
-    url += `?term=${searchTerm}`;
-  }
-
-  const featuresRequest = axios.get(url).then(
-    (response) => {
-      const featuresData = response.data;
-      return featuresData;
-    }
-  );
-
-  let preparedFeatures = null;
-  let hitFeatureCountThreshold = false;
-
-  preparedFeatures = featuresRequest.then((featuresData) => {
-
-    setFeatures((prev) => {
-
-      if (prev.length > 500) {
-        hitFeatureCountThreshold = true;
-      }
-
-      return !scrollId ? featuresData.results : prev.concat(featuresData.results);
-    });
-
-    return [featuresData.scroll_id, featuresData.results];
-  });
-
-  preparedFeatures.then(([ newScrollId, results ]) => {
-
-    // when there's no scroll id, we've hit the end of the results
-    if (newScrollId && !hitFeatureCountThreshold) {
-      // if we get a scroll id back, there are more results
-      // so call fetchModels again to fetch the next set
-      fetchFeatures(setFeatures, setFeaturesLoading, setFeaturesError, searchTerm, newScrollId);
-    } else {
-      setFeaturesLoading(false);
-    }
-  });
-
-  preparedFeatures.catch((error) => {
-    console.log('error:', error);
-    setFeaturesError(true);
-  });
-};
-
-/**
- *
- */
-const CustomTablePagination = props => {
-
-  const { state, apiRef } = useGridSlotComponentProps();
-
-  return (
-      <TablePagination
-                              labelDisplayedRows={({from, to, count}) => {
-                                const displayCount = count > 500 ? "Many" : count;
-                                return `${from}-${to} of ${displayCount}`;
-                              }}
-    {...props}
-    page={state.pagination.page}
-    onPageChange={(event, value) => {
-      return apiRef.current.setPage(value);
-    }}
-    rowsPerPage={100}
-    count={state.pagination.rowCount}
-      />
-  );
-};
-
-/**
- *
- */
-function CustomLoadingOverlay() {
-  return (
-    <GridOverlay>
-      <div style={{ position: 'absolute', top: 0, width: '100%' }}>
-        <LinearProgress />
-      </div>
-    </GridOverlay>
-  );
-}
 
 /**
  *
@@ -321,8 +170,6 @@ function ViewDatasets() {
   const [displayDeprecated, setDisplayDeprecated] = useState(false);
   const [displayedDatasets, setDisplayedDatasets] = useState([]);
 
-  // TODO probably move the features grid and related logic to its own file
-
   const [mode, setMode] = useState("datasets");
 
   const handleMode = (event, newMode) => {
@@ -331,20 +178,6 @@ function ViewDatasets() {
 
   const isModeDatasets = mode === "datasets";
   const isModeFeatures = mode === "features";
-
-  const [featuresSearchTerm, setFeaturesSearchTerm] = useState('');
-  const [featuresSearchTermValue, setFeaturesSearchTermValue] = useState('');
-
-  const updateFeaturesSearchTerm = useCallback(debounce(setFeaturesSearchTerm, 500), []);
-
-  const [features, setFeatures] = useState([]);
-  const [featuresError, setFeaturesError] = useState(false);
-  const [featuresLoading, setFeaturesLoading] = useState(false); // Loading unused for now.
-
-
-  useEffect(() => {
-      fetchFeatures(setFeatures, setFeaturesLoading, setFeaturesError, featuresSearchTerm);
-  }, [featuresSearchTerm]);
 
   useEffect(() => {
     // only do this once when the page loads
@@ -360,21 +193,6 @@ function ViewDatasets() {
       setDisplayedDatasets(filtered);
     }
   }, [datasets]);
-
-  const clearFeaturesSearch = () => {
-    setFeaturesSearchTerm('');
-    setFeaturesSearchTermValue('');
-  };
-
-  const handleFeatureSearchChange = ({ target: { value } }) => {
-    // if we have no search term, clear everything
-    if (value.length === 0) {
-      clearFeaturesSearch();
-      return;
-    }
-    setFeaturesSearchTermValue(value);
-    updateFeaturesSearchTerm(value);
-  };
 
   if (datasetsLoading) {
     return <LoadingOverlay text="Loading Datasets" />;
@@ -410,125 +228,79 @@ function ViewDatasets() {
       maxWidth="xl"
     >
       <div style={{display: "flex", flexDirection: "column", width: "100%", flex: "1 0 100%"}}>
-      <Typography
-        className={classes.header}
-        component="h3"
-        variant="h4"
-        align="center"
-      >
-        All Datasets & Features
-      </Typography>
-
-      <div style={{display: "flex", alignItems: "center"}}>
-        <Typography variant="h6">
-          Browse By
-        </Typography>
-        &nbsp;
-        <ToggleButtonGroup
-          exclusive
-          value={mode}
-          onChange={handleMode}
+        <Typography
+          className={classes.header}
+          component="h3"
+          variant="h4"
+          align="center"
         >
-          <ToggleButton value="datasets">
-            Datasets
-          </ToggleButton>
-          <ToggleButton value="features">
-            Features
-          </ToggleButton>
-        </ToggleButtonGroup>
-      </div>
+          All Datasets & Features
+        </Typography>
 
-      <br />
-
-        <div className={classes.gridContainer} style={{flex: 1, display: "flex", flexDirection: "column"}}>
-        {isModeDatasets ? (
-          <>
-            <div className={classes.aboveTableWrapper}>
-              <SearchDatasets
-                setSearchedDatasets={setSearchedDatasets}
-                datasets={displayedDatasets}
-              />
-              <Button
-                component={Link}
-                size="large"
-                variant="outlined"
-                color="primary"
-                disableElevation
-                to="/datasets/register"
-              >
-                Register a New Dataset
-              </Button>
-              <Button
-                color="primary"
-                disableElevation
-                variant="outlined"
-                size="large"
-                onClick={toggleDeprecatedDatasets}
-              >
-                {displayDeprecated ? 'Hide Deprecated Datasets' : 'Show Deprecated Datasets'}
-              </Button>
-            </div>
-
-            <DataGrid
-              autoHeight
-              columns={columns}
-              rows={searchedDatasets !== null ? searchedDatasets : displayedDatasets}
-              getRowClassName={
-                (params) => params.getValue(params.id, 'deprecated') && 'deprecatedDataset'
-              }
-            />
-          </>
-        ) : featuresError ? (
-          <Typography>
-            Error loading features.
+        <div style={{display: "flex", alignItems: "center"}}>
+          <Typography variant="h6">
+            Browse By
           </Typography>
-        ) : (
-          <div style={{flex: 1, display: "flex", flexDirection: "column"}}>
-            <div className={classes.aboveTableWrapper}>
-              <TextField
-                style={{
-                  width: '400px',
-                  marginRight: 16,
-                }}
-                label="Filter Features"
-                variant="outlined"
-                value={featuresSearchTermValue}
-                onChange={handleFeatureSearchChange}
-                role="searchbox"
-                InputProps={{
-                  endAdornment: (
-                    <InputAdornment position="end">
-                      <IconButton onClick={clearFeaturesSearch}><CancelIcon /></IconButton>
-                    </InputAdornment>
-                  )
-                }}
+          &nbsp;
+          <ToggleButtonGroup
+            exclusive
+            value={mode}
+            onChange={handleMode}
+          >
+            <ToggleButton value="datasets">
+              Datasets
+            </ToggleButton>
+            <ToggleButton value="features">
+              Features
+            </ToggleButton>
+          </ToggleButtonGroup>
+        </div>
+
+        <br />
+
+        <div className={classes.gridContainer}>
+          {isModeDatasets ? (
+            <>
+              <div className={classes.aboveTableWrapper}>
+                <SearchDatasets
+                  setSearchedDatasets={setSearchedDatasets}
+                  datasets={displayedDatasets}
+                />
+                <Button
+                  component={Link}
+                  size="large"
+                  variant="outlined"
+                  color="primary"
+                  disableElevation
+                  to="/datasets/register"
+                >
+                  Register a New Dataset
+                </Button>
+                <Button
+                  color="primary"
+                  disableElevation
+                  variant="outlined"
+                  size="large"
+                  onClick={toggleDeprecatedDatasets}
+                >
+                  {displayDeprecated ? 'Hide Deprecated Datasets' : 'Show Deprecated Datasets'}
+                </Button>
+              </div>
+
+              <DataGrid
+                autoHeight
+                columns={columns}
+                rows={searchedDatasets !== null ? searchedDatasets : displayedDatasets}
+                getRowClassName={
+                  (params) => params.getValue(params.id, 'deprecated') && 'deprecatedDataset'
+                }
               />
-            </div>
-            <Alert
-              variant="outlined"
-              severity="info"
-              style={{border: "none"}}
-            >
-              Click on a row, then CTRL+C or CMD+C to copy contents.
-            </Alert>
-            <DataGrid
-              autoHeight
-              classes={{
-                root: classes.featureGridRoot
-              }}
-              components={{
-                LoadingOverlay: CustomLoadingOverlay,
-                Pagination: CustomTablePagination
-              }}
-              loading={featuresLoading}
-              getRowId={(row) => `${row.owner_dataset.id}-${row.name}`}
-              columns={featureColumns}
-              rows={features}
-            />
-          </div>
-        )}
+            </>
+          ) : (
+            <ViewFeatures />
+          )}
+        </div>
       </div>
-    </div>
     </Container>
   );
 }
